+++ CHANGES.md
@@ -2,3 +2,5 @@ Next release
 ============
 
 - `Py.import` is an alias for `Py.Import.import_module`.
+- Use `*_opt` naming convention for the functions that return an option
+  instead of an exception: `Py.import_opt`, `Py.Object.find_opt`,...